collapse getStopPolyline into getStopRouteByIndex
The trekker had a thin wrapper file delegating to the shared
`getStopPolyline` — same return shape, same semantics, two names
for one concept. Move the canonical implementation to
`getStopRouteByIndex` in tourGeometry, accept `Maybe<Tour>` (so
the wrapper's null guard moves up), and delete the wrapper file.

Three call sites updated to import from `@/shared/tourGeometry`:
MapboxLocationSelector, getFullTourRoute, TourMap.

Tests: the wrapper had 5 tests largely overlapping tourGeometry's
coverage; folded the unique "null tour" case into the shared
suite and dropped the rest.
